% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bow.R
\name{bow}
\alias{bow}
\alias{is.polite}
\title{Introduce yourself to the host}
\usage{
bow(
  url,
  user_agent = "polite R package",
  delay = 5,
  times = 3,
  force = FALSE,
  verbose = FALSE,
  ...
)

is.polite(x)
}
\arguments{
\item{url}{URL}

\item{user_agent}{character value passed to user agent string}

\item{delay}{desired delay between scraping attempts. Final value will be the maximum of desired and mandated delay, as stipulated by \code{robots.txt} for relevant user agent}

\item{times}{number of times to attempt scraping. Default is 3.}

\item{force}{refresh all memoised functions. Clears up \code{robotstxt} and \code{scrape} caches. Default is \code{FALSE}}

\item{verbose}{TRUE/FALSE}

\item{...}{other curl parameters wrapped into \code{httr::config} function}

\item{x}{object of class \code{polite}, \code{session}}
}
\value{
object of class \code{polite}, \code{session}
}
\description{
Introduce yourself to the host
}
\examples{
\donttest{
 library(polite)

 host <- "https://www.wikipedia.org"
 session <- bow(host)
 session
}
}
