% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nod.R
\name{nod}
\alias{nod}
\title{Agree modification of session path with the host}
\usage{
nod(bow, path, verbose = FALSE)
}
\arguments{
\item{bow}{object of class \code{polite}, \code{session} created by \code{polite::bow()}}

\item{path}{string value of path/URL to follow. The function accepts either a path (string part of URL following domain name) or a full URL}

\item{verbose}{\code{TRUE}/\code{FALSE}}
}
\value{
object of class \code{polite}, \code{session} with modified URL
}
\description{
Agree modification of session path with the host
}
\examples{
\donttest{
 library(polite)

 host <- "https://www.wikipedia.org"
 session <- bow(host) \%>\%
              nod(path="wiki")
 session
}
}
