% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%otherwise\%}
\title{Null coalescing operator}
\usage{
lhs \%otherwise\% rhs
}
\description{
See \code{purrr::\link[purrr]{\%||\%}} for details.
}
\keyword{internal}
