% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bow.R
\name{bow}
\alias{bow}
\title{Inroduce yourself to the host}
\usage{
bow(url, user_agent = "polite R package - https://github.com/dmi3kno/polite",
  force = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{url}{url}

\item{user_agent}{character value passed to user_agent string}

\item{force}{refresh all memoised functions. Clears up all robotstxt and scrape cache. Default is FALSE.}

\item{verbose}{TRUE/FALSE}

\item{...}{other curl parameters wrapped into httr::config function}
}
\value{
object of class \code{polite}, \code{session}
}
\description{
Inroduce yourself to the host
}
\examples{
\dontrun{
 library(polite)

 host <- "https://www.cheese.com"
 session <- bow(host)
 session
}
}
