% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bow.R
\name{bow}
\alias{bow}
\alias{is.polite}
\title{Inroduce yourself to the host}
\usage{
bow(url,
  user_agent = "polite R package - https://github.com/dmi3kno/polite",
  delay = 5, force = FALSE, verbose = FALSE, ...)

is.polite(x)
}
\arguments{
\item{url}{url}

\item{user_agent}{character value passed to user_agent string}

\item{delay}{desired delay between scraping attempts. Final value will be the maximum of desired and mandated delay, as stipulated by robots.txt for relevant user agent}

\item{force}{refresh all memoised functions. Clears up all robotstxt and scrape cache. Default is FALSE.}

\item{verbose}{TRUE/FALSE}

\item{...}{other curl parameters wrapped into httr::config function}

\item{x}{object of class \code{polite session}}
}
\value{
object of class \code{polite}, \code{session}
}
\description{
Inroduce yourself to the host
}
\examples{
\dontrun{
 library(polite)

 host <- "https://www.cheese.com"
 session <- bow(host)
 session
}
}
