% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_basename}
\alias{guess_basename}
\title{Guess download file name from the URL}
\usage{
guess_basename(x)
}
\arguments{
\item{x}{url to guess basename from}
}
\value{
guessed file name
}
\description{
Guess download file name from the URL
}
\examples{
guess_basename("https://bit.ly/polite_sticker")

}
