% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nod.R
\name{nod}
\alias{nod}
\title{Agree modification of session path with the host}
\usage{
nod(bow, path)
}
\arguments{
\item{bow}{object of class \code{polite}, \code{session} created by \code{polite::bow()}}

\item{path}{string value of path/url to follow. The function accepts both path (string part of url followin domain name) or a full url.}
}
\value{
object of class \code{polite}, \code{session} with modified url
}
\description{
Agree modification of session path with the host
}
\examples{
\dontrun{
 library(polite)

 host <- "https://www.cheese.com"
 session <- bow(host) \%>\%
              nod(path="by_type")
 session
}
}
