% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bow.R
\name{print.polite}
\alias{print.polite}
\title{Print host introduction object}
\usage{
\method{print}{polite}(x, ...)
}
\arguments{
\item{x}{object of class \code{polite}, \code{session}}

\item{...}{other parameters passed to methods}
}
\description{
Print host introduction object
}
