% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rip.R
\name{rip}
\alias{rip}
\title{Polite file download}
\usage{
rip(bow, suffix = NULL, sep = "__", path = "downloads",
  overwrite = FALSE, mode = "wb", ...)
}
\arguments{
\item{bow}{host introduction object of class polite, session created by bow() or nod}

\item{suffix}{optional characters added to file name}

\item{sep}{separator between file name and suffix. Default "__"}

\item{path}{path where file should be saved. Defults to folder named "downloads" created in the working directory}

\item{overwrite}{if TRUE will overwrite file on disk}

\item{mode}{character. The mode with which to write the file. Useful values are "w", "wb" (binary), "a" (append) and "ab". Not used for methods "wget" and "curl".}

\item{...}{other parameters passed to download.file}
}
\value{
Full path to file indicated by url saved on disk
}
\description{
Polite file download
}
\examples{
\dontrun{
bow("www.mysite.com") \%>\%
  nod("file.txt") \%>\%
  rip()
}
}
