% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bow.R
\name{set_scrape_delay}
\alias{set_scrape_delay}
\alias{set_rip_delay}
\title{Reset scraping/ripping rate limit}
\usage{
set_scrape_delay(delay)

set_rip_delay(delay)
}
\arguments{
\item{delay}{Delay between subsequent requests. Default for package is 5 sec.
It can be set lower only under the condition of specifying a custom user-agent string.}
}
\value{
Updates rate-limit property of \code{scrape} and \code{rip} functions, respectively.
}
\description{
Reset scraping/ripping rate limit
}
\examples{
\donttest{
 library(polite)

 host <- "https://www.cheese.com"
 session <- bow(host)
 session
}
}
