% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-manners.R
\name{use_manners}
\alias{use_manners}
\title{Use manners in your own package or script}
\usage{
use_manners(save_as = "R/polite-scrape.R", open = TRUE)
}
\arguments{
\item{save_as}{File where function should be created Defaults to "\code{R/polite-scrape.R}"}

\item{open}{if \code{TRUE}, open the resultant files}
}
\description{
Creates collection of  \code{polite} functions for scraping and downloading
}
